requirements.txt
attrs==22.2.0
blinker==1.5
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-Session==0.4.0
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mock==5.0.1
mysql-connector-python==8.0.32
mysqlclient==2.1.1
packaging==23.0
pluggy==1.0.0
protobuf==3.20.3
pytest==7.2.1
requests==2.28.2
responses==0.22.0
toml==0.10.2
tomli==2.0.1
types-toml==0.10.8.2
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.12.0
